reimplement lighting settings as component
