+++ src/components.rs
@@ -16,7 +16,23 @@ impl Default for AmbientLight2d {
     fn default() -> Self {
         Self {
             color: Color::WHITE,
-            brightness: 0.8,
+            brightness: 1.0,
+        }
+    }
+}
+
+#[derive(Component, Clone, Reflect)]
+pub struct Lighting2dSettings {
+    pub shadow_softness: f32,
+    /// if false the shadow softness is calculated in relation to the viewport size
+    pub fixed_resolution: bool,
+}
+
+impl Default for Lighting2dSettings {
+    fn default() -> Self {
+        Self {
+            shadow_softness: 0.0,
+            fixed_resolution: true,
         }
     }
 }
