+++ src/lib.rs
@@ -3,10 +3,8 @@ mod extract;
 mod pipeline;
 mod plugin;
 mod prepare;
-mod resources;
 
 pub mod prelude {
     pub use crate::components::*;
     pub use crate::plugin::*;
-    pub use crate::resources::*;
 }
